Remake subtask models, subtasks now addable/removable
Split the task models file into task models and saveable task models
(updated for serialization). Subtasks now cannot be recursive
and don't have an id. They only have name, description and
whether they are completed. More is in the future.

Subtasks can be added and removed in the task edit window.
Next is editing them via their own window and marking them
completed. Also, invariant with the parent task needs to be held:
if task is completed, all of its subtasks are completed too.
This needs to go both ways (the other part has to adapt).